// src/theme/Models/type.colors.d.ts
export interface PropsColorsCore {
  PRYMARY_BASE_01: string;
  PRYMARY_BASE_02: string;
  SECONDARY_BASE_01: string;
  SECONDARY_BASE_02: string;
  SECONDARY_BASE_03: string;
  SECONDARY_BASE_04: string;
  SECONDARY_BASE_05: string;
  SECONDARY_BASE_06: string;
  LIGHT_BASE_01: string;
  LIGHT_BASE_02: string;
  LIGHT_BASE_03: string;
  LIGHT_BASE_04: string;
  LIGHT_BASE_05: string;
  DARK_BASE_01: string;
  DARK_BASE_02: string;
  DARK_BASE_03: string;
  DARK_BASE_04: string;
  DARK_BASE_05: string;
  SUCCESS_BASE_01: string;
  SUCCESS_BASE_02: string;
  SUCCESS_BASE_03: string;
  SUCCESS_BASE_04: string;
  SUCCESS_BASE_05: string;
  SUCCESS_BASE_06: string;
  ERROR_BASE_01: string;
  ERROR_BASE_02: string;
  ERROR_BASE_03: string;
  ERROR_BASE_04: string;
  ERROR_BASE_05: string;
  ALERT_COLOR_01: string;
  ALERT_COLOR_02: string;
  ALERT_COLOR_03: string;
  ALERT_COLOR_04: string;
  ALERT_COLOR_05: string;
  HIGHLIGHT_COLOR_01: string;
  HIGHLIGHT_COLOR_02: string;
  HIGHLIGHT_COLOR_03: string;
  HIGHLIGHT_COLOR_04: string;
  HIGHLIGHT_COLOR_05: string;
  SAFE_COLOR_01: string;
  SAFE_COLOR_02: string;
  SAFE_COLOR_03: string;
  SAFE_COLOR_04: string;
}

export interface PropsColors extends PropsColorsCore {}

export type PropsTokenColorsCore =
  | "PRYMARY_BASE_01"
  | "PRYMARY_BASE_02"
  | "SECONDARY_BASE_01"
  | "SECONDARY_BASE_02"
  | "SECONDARY_BASE_03"
  | "SECONDARY_BASE_04"
  | "SECONDARY_BASE_05"
  | "SECONDARY_BASE_06"
  | "LIGHT_BASE_01"
  | "LIGHT_BASE_02"
  | "LIGHT_BASE_03"
  | "LIGHT_BASE_04"
  | "LIGHT_BASE_05"
  | "DARK_BASE_01"
  | "DARK_BASE_02"
  | "DARK_BASE_03"
  | "DARK_BASE_04"
  | "DARK_BASE_05"
  | "SUCCESS_BASE_01"
  | "SUCCESS_BASE_02"
  | "SUCCESS_BASE_03"
  | "SUCCESS_BASE_04"
  | "SUCCESS_BASE_05"
  | "SUCCESS_BASE_06"
  | "ERROR_BASE_01"
  | "ERROR_BASE_02"
  | "ERROR_BASE_03"
  | "ERROR_BASE_04"
  | "ERROR_BASE_05"
  | "ALERT_COLOR_01"
  | "ALERT_COLOR_02"
  | "ALERT_COLOR_03"
  | "ALERT_COLOR_04"
  | "ALERT_COLOR_05"
  | "HIGHLIGHT_COLOR_01"
  | "HIGHLIGHT_COLOR_02"
  | "HIGHLIGHT_COLOR_03"
  | "HIGHLIGHT_COLOR_04"
  | "HIGHLIGHT_COLOR_05"
  | "SAFE_COLOR_01"
  | "SAFE_COLOR_02"
  | "SAFE_COLOR_03"
  | "SAFE_COLOR_04";